pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A short description of the project"
authors = ["vikas (vikas@neoito.com)"]

[tool.poetry.dependencies]
python = "^3.8.10"
requests = "2.28.2"
loguru = "^0.7.0"
joblib = "^1.2.0"
sqlalchemy = "1.4.44"
openai = "0.27.2"
langchain = "^0.0.151"
typing-extensions = "4.5.0"
typing-inspect = "0.8.0"
python-multipart = "0.0.6"
numpy = "1.24.2"
aiohttp = "3.8.4"
marshmallow-enum = "1.5.1"
marshmallow = "3.19.0"
python-dotenv = "0.21.0"
fastapi = "0.95.0"
pydantic = "1.10.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
psycopg2-binary = "2.9.5"
packaging = "23.0"
starlette = "0.26.1"
async-timeout = "4.0.2"


[tool.poetry.group.dev.dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^2.15.8"


[tool.poetry.scripts]
start = "app.main:start"


[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.inipsycopg2-binary
  |Dockerfile
  |Jenkinfile
)/
'''